# Improved performance of searching for the next `Selectable` in scenes with a large number of root objects