--- a/Assets/EasyTab/Source/Drivers/IEasyTabDriver.cs
+++ b/Assets/EasyTab/Source/Drivers/IEasyTabDriver.cs
@@ -1,3 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+
 namespace EasyTab
 {
     public interface IEasyTabDriver