--- a/Assets/Tests/Source/PerformanceTests.cs.meta
+++ b/Assets/Tests/Source/PerformanceTests.cs.meta
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:dfb436b64090ab7d19526f961bb4d03eed168f65756e21d87e0d5d1b6b090f2d
+size 83