Improving the performance of the Navigate method depending on the number of root objects:

| N | old time of `Navigate` | new time of `Navigate` |
|---|----------------------|---|
| 20 | 0,1 ms |  0,02 ms |
| 250 | 1,6 ms | 0,2 ms |
| 1000 | 20,8 ms | 1,0 ms |
| 5000 | 455,0 ms | 5,1 ms |
| 20000 | 6973,7 ms | 20,2 ms |